Improve modeling of hill lists
The hill lists tracked by the app may well change (e.g. several changes
were made in version 5.2) and modeling these using an enum would mean
breaking the public API if cases were added, removed or renamed.

Lists are now modeled using a struct, and additionally have a property
to store the flag displayed for the list in the app's UI. Storing the
flag has the added benefit of allowing the content of lists in Ireland
and Northern Ireland to be inferred (see the API documentation for the
`HillList` struct for details).